# Conflicts:
#	src/main/java/com/example/carsharing/CarSharingApplication.java
#	src/main/java/com/example/carsharing/service/impl/RentalServiceImpl.java

+++ src/main/java/com/example/carsharing/strategy/impl/PaymentTotalAmountHandler.java
@@ -11,7 +11,7 @@ public class PaymentTotalAmountHandler implements TotalAmountHandler {
 
     @Override
     public BigDecimal getTotalAmount(Rental rental, BigDecimal dailyFee) {
-        BigDecimal rentDays = BigDecimal.valueOf(getRentDays(rental));
+        BigDecimal rentDays = getRentDays(rental);
         return rentDays.multiply(dailyFee);
     }
